Save, but don't print, state refresh output
Some customers have been hitting character limits on CircleCI stdout. This allows us to save the entire output of a plan, but not display the extremely verbose step of refreshing all resources. As customers move to layers, this will be less of an issue, but this helps them make the move.