Use a single repository template
Just use a single repository template, instead of different ones for the common and single-account repositories. The major change here was to ensure the linting step happens for `modules/*` and `layers/*`, and that we don't plan or apply when there's no `layers` directory (we exit).

The other implication is that we'll have a hold (approval) step and an apply step, for the CI in the common repo. However, we're going to remove the hold step soon anyway, and merge on master.